Add optional setting to lock chests on open
To enable, run /data remove storage ironman:settings popChests
Just an extra option for if you think instamining chests is too cheaty, maybe this will be the default later? Either way it's in the datapack now. Should do this for any block you can expect to have loot in it in the natural world, every other block behaves the same way it used to.